refactor(T10064): decompose task-ops.ts into 10 single-concern files behind barrel
3408 LOC → 10 files ≤780 LOC each. Zero behavior change verified via
characterization tests + barrel re-export preserves all import sites.

Files created:
- task-tree.ts      (305 LOC) — buildTaskTree + hierarchy helpers
- task-next.ts      (157 LOC) — coreTaskNext
- task-blockers.ts  (120 LOC) — coreTaskBlockers
- task-analyze.ts   (261 LOC) — coreTaskAnalyze + complexity helpers
- task-reparent.ts  (464 LOC) — restore/reopen/reorder/reparent/unarchive
- task-data.ts      (498 LOC) — deps/relates/stats/depends query ops
- task-import.ts    (577 LOC) — import/export/history/lint/validate
- tree-render.ts    (198 LOC) — text/mermaid dep-tree renderers
- engine-wrap.ts    (576 LOC) — core EngineResult wrappers + coreTaskCancel
- engine-wrap-ops.ts (416 LOC) — data/query EngineResult wrappers

task-ops.ts becomes a 103-LOC barrel re-export preserving all 60+ import sites.

Part of Saga T9831 / Epic T9834 E-CORE-DECOMP.
